create IDentityIO interface
