+++ identity/src/main/java/com/stripe/android/identity/navigation/IdentityCameraScanFragment.kt
@@ -105,7 +105,6 @@ internal abstract class IdentityCameraScanFragment(
                 onSuccess = {
                     identityScanViewModel.uploadResult(
                         finalResult,
-                        requireContext(),
                         it.documentCapture
                     )
                 },
