+++ identity/src/main/java/com/stripe/android/identity/navigation/IdentityFragmentFactory.kt
@@ -8,6 +8,8 @@ import com.stripe.android.camera.CameraPermissionEnsureable
 import com.stripe.android.identity.IdentityVerificationSheetContract
 import com.stripe.android.identity.VerificationFlowFinishable
 import com.stripe.android.identity.networking.DefaultIdentityRepository
+import com.stripe.android.identity.utils.DefaultIdentityIO
+import com.stripe.android.identity.utils.IdentityIO
 import com.stripe.android.identity.viewmodel.FrontBackUploadViewModel
 import com.stripe.android.identity.viewmodel.IdentityScanViewModel
 import com.stripe.android.identity.viewmodel.IdentityViewModel
@@ -15,6 +17,8 @@ import com.stripe.android.identity.viewmodel.PassportUploadViewModel
 
 /**
  * Factory for creating Identity fragments.
+ *
+ * TODO(ccen) Daggerize the dependencies
  */
 internal class IdentityFragmentFactory(
     context: Context,
@@ -23,21 +27,26 @@ internal class IdentityFragmentFactory(
     verificationArgs: IdentityVerificationSheetContract.Args,
     private val verificationFlowFinishable: VerificationFlowFinishable
 ) : FragmentFactory() {
-    private val identityRepository = DefaultIdentityRepository(context)
+    private val identityIO: IdentityIO = DefaultIdentityIO(context)
+    private val identityRepository =
+        DefaultIdentityRepository(context = context, identityIO = identityIO)
     private val identityScanViewModelFactory =
         IdentityScanViewModel.IdentityScanViewModelFactory(
             identityRepository,
-            verificationArgs
+            verificationArgs,
+            identityIO
         )
     private val frontBackUploadViewModelFactory =
         FrontBackUploadViewModel.FrontBackUploadViewModelFactory(
             identityRepository,
-            verificationArgs
+            verificationArgs,
+            identityIO
         )
     private val passportUploadViewModelFactory =
         PassportUploadViewModel.PassportUploadViewModelFactory(
             identityRepository,
-            verificationArgs
+            verificationArgs,
+            identityIO
         )
 
     internal val identityViewModelFactory = IdentityViewModel.IdentityViewModelFactory(
