+++ identity/src/main/java/com/stripe/android/identity/navigation/PassportUploadFragment.kt
@@ -122,7 +122,6 @@ internal class PassportUploadFragment(
             onSuccess = { verificationPage ->
                 passportUploadViewModel.uploadImage(
                     passport,
-                    requireContext(),
                     verificationPage.documentCapture,
                     uploadMethod
                 )
