+++ identity/src/test/java/com/stripe/android/identity/navigation/DriverLicenseScanFragmentTest.kt
@@ -337,7 +337,6 @@ internal class DriverLicenseScanFragmentTest {
         successCaptor.lastValue.invoke(mockVerificationPage)
         verify(mockIdentityScanViewModel).uploadResult(
             same(finalResult),
-            eq(context),
             same(mockDocumentCapturePage)
         )
     }
