+++ identity/src/test/java/com/stripe/android/identity/networking/DefaultIdentityRepositoryTest.kt
@@ -20,6 +20,7 @@ import com.stripe.android.identity.networking.models.CollectedDataParam.Companio
 import com.stripe.android.identity.networking.models.ConsentParam
 import com.stripe.android.identity.networking.models.VerificationPage
 import com.stripe.android.identity.networking.models.VerificationPageData
+import com.stripe.android.identity.utils.IdentityIO
 import kotlinx.coroutines.runBlocking
 import org.json.JSONObject
 import org.junit.Test
@@ -38,11 +39,14 @@ import kotlin.test.assertFailsWith
 
 @RunWith(RobolectricTestRunner::class)
 class DefaultIdentityRepositoryTest {
-
+    private val mockIO = mock<IdentityIO>().also {
+        whenever(it.createTFLiteFile()).thenReturn(mock())
+    }
     private val mockStripeNetworkClient: StripeNetworkClient = mock()
     private val identityRepository = DefaultIdentityRepository(
         ApplicationProvider.getApplicationContext(),
-        mockStripeNetworkClient
+        mockStripeNetworkClient,
+        mockIO
     )
 
     private val requestCaptor: KArgumentCaptor<StripeRequest> = argumentCaptor()
